@@ -2,7 +2,7 @@
  * rflutter_alert
  * Created by Ratel
  * https://ratel.com.tr
- * 
+ *
  * Copyright (c) 2018 Ratel, LLC. All rights reserved.
  * See LICENSE for distribution and usage details.
  */
@@ -19,33 +19,33 @@ import 'dialog_button.dart';
 ///
 /// You must call the "show()" method to view the alert class you have defined.
 class Alert {
-  final String id;
+  final String? id;
   final BuildContext context;
-  final AlertType type;
+  final AlertType? type;
   final AlertStyle style;
-  final Widget image;
+  final Widget? image;
   final String title;
-  final String desc;
+  final String? desc;
   final Widget content;
-  final List<DialogButton> buttons;
-  final Function closeFunction;
-  final Icon closeIcon;
+  final List<DialogButton>? buttons;
+  final Function? closeFunction;
+  final Widget? closeIcon;
   final bool onWillPopActive;
-  final AlertAnimation alertAnimation;
   final bool useRootNavigator;
+  final AlertAnimation? alertAnimation;
 
   /// Alert constructor
   ///
   /// [context], [title] are required.
   Alert({
-    @required this.context,
+    required this.context,
     this.id,
     this.type,
     this.style = const AlertStyle(),
     this.image,
-    @required this.title,
+    required this.title,
     this.desc,
-    this.content,
+    this.content = const SizedBox(),
     this.buttons,
     this.closeFunction,
     this.closeIcon,
@@ -55,16 +55,14 @@ class Alert {
   });
 
   /// Displays defined alert window
-  Future<bool> show() async {
+  Future<bool?> show() async {
     return await showGeneralDialog(
         context: context,
-        pageBuilder: (BuildContext buildContext, Animation<double> animation,
-            Animation<double> secondaryAnimation) {
+        pageBuilder: (BuildContext buildContext, Animation<double> animation, Animation<double> secondaryAnimation) {
           return _buildDialog();
         },
         barrierDismissible: style.isOverlayTapDismiss,
-        barrierLabel:
-            MaterialLocalizations.of(context).modalBarrierDismissLabel,
+        barrierLabel: MaterialLocalizations.of(context).modalBarrierDismissLabel,
         barrierColor: style.overlayColor,
         useRootNavigator: useRootNavigator,
         transitionDuration: style.animationDuration,
@@ -76,8 +74,7 @@ class Alert {
         ) =>
             alertAnimation == null
                 ? _showAnimation(animation, secondaryAnimation, child)
-                : alertAnimation(
-                    context, animation, secondaryAnimation, child));
+                : alertAnimation!(context, animation, secondaryAnimation, child));
   }
 
   // Will be added in next version.
@@ -88,16 +85,13 @@ class Alert {
   // Alert dialog content widget
   Widget _buildDialog() {
     final Widget _child = ConstrainedBox(
-      constraints: style.constraints ??
-          BoxConstraints.expand(
-              width: double.infinity, height: double.infinity),
+      constraints: style.constraints ?? BoxConstraints.expand(width: double.infinity, height: double.infinity),
       child: Align(
         alignment: style.alertAlignment,
         child: SingleChildScrollView(
           child: AlertDialog(
-              key: Key(id),
-              backgroundColor: style.backgroundColor ??
-                  Theme.of(context).dialogBackgroundColor,
+              key: id == null ? null : Key(id!),
+              backgroundColor: style.backgroundColor ?? Theme.of(context).dialogBackgroundColor,
               shape: style.alertBorder ?? _defaultShape(),
               insetPadding: style.alertPadding,
               elevation: style.alertElevation,
@@ -109,8 +103,7 @@ class Alert {
                     children: <Widget>[
                       _getCloseButton(),
                       Padding(
-                        padding: EdgeInsets.fromLTRB(
-                            20, (style.isCloseButton ? 0 : 10), 20, 0),
+                        padding: EdgeInsets.fromLTRB(20, (style.isCloseButton ? 0 : 10), 20, 0),
                         child: Column(
                           children: <Widget>[
                             _getImage(),
@@ -128,11 +121,11 @@ class Alert {
                             desc == null
                                 ? Container()
                                 : Text(
-                                    desc,
+                                    desc!,
                                     style: style.descStyle,
                                     textAlign: style.descTextAlign,
                                   ),
-                            content == null ? Container() : content,
+                            content,
                           ],
                         ),
                       )
@@ -155,9 +148,7 @@ class Alert {
         ),
       ),
     );
-    return onWillPopActive
-        ? WillPopScope(onWillPop: () async => false, child: _child)
-        : _child;
+    return onWillPopActive ? WillPopScope(onWillPop: () async => false, child: _child) : _child;
   }
 
 // Returns the close button on the top right
@@ -170,7 +161,7 @@ class Alert {
                 if (closeFunction == null) {
                   Navigator.of(context, rootNavigator: useRootNavigator).pop();
                 } else {
-                  closeFunction();
+                  closeFunction!();
                 }
               },
               child: Container(
@@ -200,14 +191,13 @@ class Alert {
     List<Widget> expandedButtons = [];
     if (style.isButtonVisible) {
       if (buttons != null) {
-        buttons.forEach(
+        buttons!.forEach(
           (button) {
             var buttonWidget = Padding(
               padding: const EdgeInsets.only(left: 2, right: 2),
               child: button,
             );
-            if ((button.width != null && buttons.length == 1) ||
-                style.buttonsDirection == ButtonsDirection.column) {
+            if ((button.width != null && buttons!.length == 1) || style.buttonsDirection == ButtonsDirection.column) {
               expandedButtons.add(buttonWidget);
             } else {
               expandedButtons.add(Expanded(
@@ -274,6 +264,7 @@ class Alert {
         );
         break;
       case AlertType.none:
+      default:
         response = Container();
         break;
     }
@@ -284,27 +275,17 @@ class Alert {
   _showAnimation(animation, secondaryAnimation, child) {
     switch (style.animationType) {
       case AnimationType.fromRight:
-        return AnimationTransition.fromRight(
-            animation, secondaryAnimation, child);
-        break;
+        return AnimationTransition.fromRight(animation, secondaryAnimation, child);
       case AnimationType.fromLeft:
-        return AnimationTransition.fromLeft(
-            animation, secondaryAnimation, child);
-        break;
+        return AnimationTransition.fromLeft(animation, secondaryAnimation, child);
       case AnimationType.fromBottom:
-        return AnimationTransition.fromBottom(
-            animation, secondaryAnimation, child);
-        break;
+        return AnimationTransition.fromBottom(animation, secondaryAnimation, child);
       case AnimationType.grow:
         return AnimationTransition.grow(animation, secondaryAnimation, child);
-        break;
       case AnimationType.shrink:
         return AnimationTransition.shrink(animation, secondaryAnimation, child);
-        break;
       case AnimationType.fromTop:
-        return AnimationTransition.fromTop(
-            animation, secondaryAnimation, child);
-        break;
+        return AnimationTransition.fromTop(animation, secondaryAnimation, child);
     }
   }
 }
